Standardize CI 20200203
