+++ composer.json
@@ -40,11 +40,30 @@
             "lib/Promise/functions.php"
         ]
     },
+    "autoload-dev": {
+        "psr-4" : {
+            "Sabre\\Event\\" : "tests/Event"
+        }
+    },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "~2.16.1",
+        "phpstan/phpstan": "^0.12",
         "phpunit/phpunit" : "^7 || ^8"
     },
-    "config" : {
-        "bin-dir" : "bin/"
+    "scripts": {
+        "phpstan": [
+            "@php vendor/bin/phpstan analyse lib tests"
+        ],
+        "cs-fixer": [
+            "@php vendor/bin/php-cs-fixer fix"
+        ],
+        "phpunit": [
+            "@php vendor/bin/phpunit --configuration tests/phpunit.xml"
+        ],
+        "test": [
+            "composer phpstan",
+            "composer cs-fixer",
+            "composer phpunit"
+        ]
     }
 }
