+++ tests/Event/ContinueCallbackTest.php
