+++ tests/Event/EmitterTest.php
