+++ tests/Event/Loop/FunctionsTest.php
