+++ tests/Event/Promise/FunctionsTest.php
