+++ tests/Event/Promise/PromiseTest.php
