+++ tests/Event/WildcardEmitterTest.php
